Aggiunge aggiornamento automatico prezzi prima degli snapshot
- Crea helper priceUpdater per aggiornare prezzi asset
- L'endpoint snapshot ora aggiorna i prezzi prima di creare lo snapshot
- Il cron job beneficia automaticamente dell'aggiornamento prezzi
- Refactoring endpoint /api/prices/update per usare helper condiviso
- Aggiunge toast di caricamento durante creazione snapshot
